requirements.txt
Flask
google-cloud-secret-manager